CrewAI/Event_Planning/requirements.txt
crewai
crewai_tools
langchain
langchain_community
python-dotenv
IPython
pydantic
openai
groq
tiktoken
requests
beautifulsoup4
litellm
